Refactor routes into individual modules
